+++ src/test/java/act/db/morphia/MorphiaDaoTestBase.java
@@ -20,18 +20,16 @@ package act.db.morphia;
  * #L%
  */
 
-import act.app.App;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.when;
+
 import act.db.Dao;
 import act.test.util.Fixture;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.osgl.$;
 
-import java.lang.reflect.Field;
 import java.util.Map;
-
-import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.when;
 
 public abstract class MorphiaDaoTestBase<T> extends MongoTestBase {