+++ src/main/java/act/db/morphia/MorphiaDao.java
@@ -20,8 +20,8 @@ package act.db.morphia;
  * #L%
  */
 
-import act.inject.param.NoBind;
 import act.util.General;
+import act.util.Stateless;
 import org.bson.types.ObjectId;
 import org.mongodb.morphia.Datastore;
 
@@ -29,7 +29,7 @@ import org.mongodb.morphia.Datastore;
  * The morphia dao base implementation use {@link org.bson.types.ObjectId} as the ID type
  */
 @General
-@NoBind
+@Stateless
 public class MorphiaDao<MODEL_TYPE>
         extends MorphiaDaoBase<ObjectId, MODEL_TYPE> {