+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # act-morphia CHANGE LOG
 
-1.6.2
+1.6.2 30/Oct/2018
+* update act to 1.8.8
+* Make MorphiaDao be stateless
 * Make MorphiaService initialization process be async #25
 * Catch up `AdaptiveRecord` updates in Act #24
 * Update `MorphiaUserBase` - catch up `UserBase` changes in act-aaa-plugin #23